==> kurstag10_Selenium1/projektbsp/01 Reading PDFs.ipynb <==
import textract
import re
import os
import pandas as pd
import progressbar
import time

#Testing with one file
text = textract.process('files/P(1)0000005.pdf', method='pdfminer')
text = str(text).replace("  ", " ")
adtext = re.search("Ad Text.*Ad Landing Page", text).group().replace("Ad Landing Page", "")
adtext = adtext.replace("Ad Text", "").replace("\\n", " ").strip()
adtext

!ls

#Creating file list
lt = os.listdir("fs")
lt.remove('.DS_Store')

lst = []
for elem in lt:
    if ".pdf" in elem:
        lst.append(elem)

#Creating dictionary for:

#'Location', 'Number', 'Date', 'Interests',
#'Age', 'Language', 'Placements', 'Ad Clicks',
#'Ad Impressions', 'Ad Spend'

bar = progressbar.ProgressBar()
results = []

for elem,i in zip(lst, bar(range(len(lst)))):
    print(elem)
    if 'ultimate_names.csv' in elem:
        continue
    
    else:
        
        text = textract.process('fs/'+elem, method='pdfminer')
        text = str(text).replace("  ", " ")
    
        if re.search("Location.*(Age)", text) != None:
            pattern = re.search("Location.*(Age)", text)
            loc = pattern.group().replace('\\nAge','').replace('Location - Living In: ','')
        else:
            loc = 'N/A'
    
        if re.search("Creation Date.*(Redactions)", str(text)) != None:
            pattern = re.search("Creation Date.*(Redactions)", str(text))
            pattern = pattern.group().replace('Creation Date\\n\\n','')
            date = re.search("\d\d/\d\d/\d\d \d\d:\d\d:\d\d [A-Z][A-Z] [A-Z]{3}", pattern)
            if date != None:
                date = date.group()
        else: 
            date = 'N/A'
    
        if re.search("Ad Clicks.*", text) != None:
            klicks = re.search("Ad Clicks.*", text)
            klicks = klicks.group()[:25].replace(",", "")
            klicks = re.search("[0-9]+", klicks)
            if klicks != None:
                klicks = klicks.group()
        else:
            klicks = 'N/A'
        
        if re.search("Ad Impressions.*", text) != None:
            imps = re.search("Ad Impressions.*", text)
            imps = imps.group()[:25].replace(",", "")
            imps = re.search("[0-9]+", imps)
            if imps != None:
                imps = imps.group()
        else:
            imps = 'N/A'
        
        if re.search("Ad Spend.*", text) != None:
            adspend = re.search("Ad Spend.*", text)
            adspend = adspend.group()[:25].replace(",", "")
            adspend = re.search("[0-9]+.[0-9]+", adspend)
            if adspend != None:
                adspend = adspend.group()
        else:
            adspend = 'N/A'  
        
        if re.search("Interests:.*", text) != None:
            interests = re.search("Interests:.*", text).group().split('\\n')[0]
        else:
            interests = 'N/A'
    
        if re.search("Age:.*", text) != None:
            age = re.search("Age:.*", text).group().split('\\n')[0]
        else:
            age = 'N/A'
        
        if re.search("Language:.*", text) != None:
            lang = re.search("Language:.*", text).group().split('\\n')[0]
        else:
            lang = 'N/A'
    
        if re.search("Placements:.*", text) != None:
            place = re.search("Placements:.*", text).group().split('\\n')[0]
        else:
            place = 'N/A'
    
        if re.search("Ad Text.*Ad Landing Page", text) != None:
            adtext = re.search("Ad Text.*Ad Landing Page", text).group().replace("Ad Landing Page", "")
            adtext = adtext.replace("Ad Text", "").replace("\\n", " ").strip()
        else:
            place = 'N/A'
        
        mini_dict = {'Location': loc,
                 'Number': elem,
                 'Date': date,
                 'Interests': interests,
                 'Age': age,
                 'Language': lang,
                 'Placements': place,
                 'Ad Text': adtext,
                 'Ad Clicks': klicks,
                 'Ad Impressions': imps,
                 'Ad Spend': adspend}
        
        results.append(mini_dict)


df = pd.DataFrame(results)

df.to_csv('readable.csv')
